+++ server/src/modules/infrastructure/custom-config/custom-config.fixture.ts
@@ -4,6 +4,9 @@ type CustomConfigEntity = {
   max_balance_single_api_key: string;
   funding_account: string;
   spot_fee: string;
+  market_making_fee: string;
+  enable_spot_fee: boolean;
+  enable_market_making_fee: boolean;
 };
 
 function createMockCustomConfigEntity(
@@ -15,6 +18,9 @@ function createMockCustomConfigEntity(
     max_balance_single_api_key: '2000',
     funding_account: 'XYZ123',
     spot_fee: '0.02',
+    market_making_fee: '0.001',
+    enable_spot_fee: true,
+    enable_market_making_fee: true,
     ...overrides,
   };
 }